More flexible/extendible method to locate pypi repository
My 'miniforge' installation on MacOs 11/Arm64 pypi has a slightly
different path to its repository. Since this may differ between
operation-systems, versions of pip and/or conda-clones, some
flexible and dynamic way to find the path may be useful.

This patch is a starting-point using the glob-module combined
with some iterating. Should be extended to work on as many
OS/pip-version/conda-clone -combos as possible